+++ lib/models/money_objects/transaction_extras/transaction_extras.dart
@@ -6,8 +6,6 @@ import 'package:money/models/money_objects/transaction_extras/transaction_extra.
 export 'package:money/models/money_objects/transaction_extras/transaction_extra.dart';
 
 class TransactionExtras extends MoneyObjects<TransactionExtra> {
-  double runningBalance = 0.00;
-
   void add(final TransactionExtra transaction) {
     transaction.id.value = iterableList().length;
     appendMoneyObject(transaction);
@@ -17,18 +15,17 @@ class TransactionExtras extends MoneyObjects<TransactionExtra> {
   List<TransactionExtra> loadFromJson(final List<MyJson> rows) {
     clear();
 
-    runningBalance = 0.00;
-
     for (final MyJson row in rows) {
       final TransactionExtra t = TransactionExtra(
         // id
+        id: row.getInt('Id'),
         // Account Id
         transaction: row.getInt('Transaction'),
         // Tax Year
         taxYear: row.getInt('TaxYear'),
         // Tax Date
-        taxDate: row.getInt('TaxDate'),
-      )..id.value = row.getInt('Id');
+        taxDate: row.getDate('TaxDate'),
+      );
 
       appendMoneyObject(t);
     }